change package name and add google-service.json
