+++ app/src/androidTest/java/com/android/mungmung/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.gomin.mungmung
+package com.android.mungmung
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
