+++ app/src/main/java/com/android/mungmung/data/UserData.kt
@@ -1,3 +1,3 @@
-package com.gomin.mungmung.data
+package com.android.mungmung.data
 
 data class UserData(val name:String, val id:String, val password:String)
