+++ app/src/main/java/com/android/mungmung/ui/NewPwFragment.kt
@@ -1,28 +1,29 @@
-package com.gomin.mungmung.ui
+package com.android.mungmung.ui
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.gomin.mungmung.databinding.FragmentChangePwBinding
+import com.android.mungmung.databinding.FragmentNewPwBinding
 
-class ChangePwFragment : Fragment() {
-    private var _binding: FragmentChangePwBinding? = null
+class NewPwFragment : Fragment() {
+
+    private var _binding: FragmentNewPwBinding? = null
     private val binding get() = _binding!!
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-
-        _binding = FragmentChangePwBinding.inflate(inflater,container,false)
+        _binding = FragmentNewPwBinding.inflate(inflater, container, false)
         return binding.root
-
     }
+
 
     override fun onDestroyView() {
         super.onDestroyView()
         _binding = null
     }
+
 }
