+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,39 +7,39 @@
 
     <fragment
         android:id="@+id/homeFragment"
-        android:name="com.gomin.mungmung.ui.HomeFragment"
+        android:name="com.android.mungmung.ui.HomeFragment"
         android:label="HomeFragment" />
     <fragment
         android:id="@+id/loginFragment"
-        android:name="com.gomin.mungmung.ui.LoginFragment"
+        android:name="com.android.mungmung.ui.LoginFragment"
         android:label="LoginFragment" />
     <fragment
         android:id="@+id/newPwFragment"
-        android:name="com.gomin.mungmung.ui.NewPwFragment"
+        android:name="com.android.mungmung.ui.NewPwFragment"
         android:label="NewPwFragment" />
     <fragment
         android:id="@+id/photoFragment"
-        android:name="com.gomin.mungmung.ui.PhotoFragment"
+        android:name="com.android.mungmung.ui.PhotoFragment"
         android:label="PhotoFragment" />
     <fragment
         android:id="@+id/profileFragment"
-        android:name="com.gomin.mungmung.ui.ProfileFragment"
+        android:name="com.android.mungmung.ui.ProfileFragment"
         android:label="ProfileFragment" />
     <fragment
         android:id="@+id/searchFragment"
-        android:name="com.gomin.mungmung.ui.SearchFragment"
+        android:name="com.android.mungmung.ui.SearchFragment"
         android:label="SearchFragment" />
     <fragment
         android:id="@+id/settingFragment"
-        android:name="com.gomin.mungmung.ui.SettingFragment"
+        android:name="com.android.mungmung.ui.SettingFragment"
         android:label="SettingFragment" />
     <fragment
         android:id="@+id/changePwFragment"
-        android:name="com.gomin.mungmung.ui.ChangePwFragment"
+        android:name="com.android.mungmung.ui.ChangePwFragment"
         android:label="fragment_change_pw"
         tools:layout="@layout/fragment_change_pw" />
     <fragment
         android:id="@+id/signupFragment"
-        android:name="com.gomin.mungmung.ui.SignupFragment"
+        android:name="com.android.mungmung.ui.SignupFragment"
         android:label="SignupFragment" />
 </navigation>
