+++ app/src/test/java/com/android/mungmung/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.gomin.mungmung
+package com.android.mungmung
 
 import org.junit.Test
