+++ build.gradle
@@ -2,7 +2,7 @@
 buildscript {
     dependencies {
         classpath 'com.android.tools.build:gradle:4.3.1'
-        classpath 'com.google.gms:google-services:4.3.15'
+        classpath 'com.google.gms:google-services:4.3.13'
     }
     repositories {
         google()
